[refactor] 상품 재고 수정
