+++ src/main/java/kr/hhplus/be/server/domain/product/entity/ProductStock.java
@@ -3,6 +3,7 @@ package kr.hhplus.be.server.domain.product.entity;
 import jakarta.persistence.Entity;
 import jakarta.persistence.Id;
 import lombok.AllArgsConstructor;
+import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 
@@ -10,6 +11,7 @@ import lombok.NoArgsConstructor;
 @Getter
 @NoArgsConstructor
 @AllArgsConstructor
+@Builder
 public class ProductStock {
 
     @Id
