+++ src/main/java/kr/hhplus/be/server/domain/product/service/ProductService.java
@@ -1,6 +1,7 @@
 package kr.hhplus.be.server.domain.product.service;
 
 import kr.hhplus.be.server.domain.product.entity.Product;
+import kr.hhplus.be.server.domain.product.entity.ProductStock;
 import kr.hhplus.be.server.domain.product.repository.ProductRepository;
 import kr.hhplus.be.server.domain.product.repository.ProductStockRepository;
 import lombok.RequiredArgsConstructor;
@@ -14,7 +15,19 @@ public class ProductService {
     private final ProductStockRepository productStockRepository;
 
     public Product getProduct(Long productId) {
-        return productRepository.findById(productId)
+        Product product = productRepository.findById(productId)
                 .orElseThrow(() -> new RuntimeException("상품을 조회할 수 없습니다."));
+        ProductStock productStock = productStockRepository.findById(productId)
+                .orElseThrow(() -> new RuntimeException("상품 재고를 조회할 수 없습니다."));
+
+        product.setProductStock(productStock);
+        return product;
+    }
+
+    public ProductStock reduceStock(Long productId, int orderQuantity) {
+        ProductStock productStock = productStockRepository.findById(productId)
+                .orElseThrow(() -> new RuntimeException("상품 재고를 조회할 수 없습니다."));
+        productStock.reduceStock(orderQuantity);
+        return productStockRepository.save(productStock);
     }
 }
